test: refactor extracting reusable logic in trait
